+++ blogpessoal/blogpessoal/src/main/java/com/generation/blogpessoal/controller/TemaController.java
@@ -1,6 +1,7 @@
 package com.generation.blogpessoal.controller;
 
 import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
@@ -15,19 +16,22 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import com.generation.blogpessoal.model.Tema;
 import com.generation.blogpessoal.repository.TemaRepository;
 
 @RestController
 @CrossOrigin(origins = "*", allowedHeaders = "*")
-@RequestMapping("/tema")
+@RequestMapping("/temas")
 public class TemaController {
 
 	@Autowired
 	private TemaRepository temaRepository;
 	
+	@GetMapping
 	public ResponseEntity<List<Tema>> getAll(){
 		return ResponseEntity.ok(temaRepository.findAll());
 	}
@@ -39,7 +43,7 @@ public class TemaController {
 				.orElse(ResponseEntity.notFound().build());
 	}
 	
-	@GetMapping("descricao/{decricao}")
+	@GetMapping("/descricao/{decricao}")
 	public ResponseEntity<List<Tema>> getByName(@PathVariable String descricao){
 		return ResponseEntity.ok(temaRepository.findAllByDescricaoContainingIgnoreCase(descricao));
 		
@@ -51,16 +55,25 @@ public class TemaController {
 				.body(temaRepository.save(tema));
 	}
 	
-	@PutMapping
-	public ResponseEntity<Tema> put (@Valid@RequestBody Tema tema){
-		return ResponseEntity.ok(temaRepository.save(tema));
-		
-	}
+    @PutMapping
+    public ResponseEntity<Tema> put(@Valid @RequestBody Tema tema){
+        return temaRepository.findById(tema.getId())
+            .map(resposta -> ResponseEntity.status(HttpStatus.CREATED)
+            .body(temaRepository.save(tema)))
+            .orElse(ResponseEntity.status(HttpStatus.NOT_FOUND).build());
+    }
 	
-	@DeleteMapping("/{id}")
-	public void delete(@PathVariable Long id ) {
-		temaRepository.deleteById(id);
-	}
+    @ResponseStatus(HttpStatus.NO_CONTENT)
+    @DeleteMapping("/{id}")
+    public void delete(@PathVariable Long id) {
+        Optional<Tema> tema = temaRepository.findById(id);
+        
+        if(tema.isEmpty())
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        
+        temaRepository.deleteById(id);              
+    }
+
 	
 	
 }
